Split cmake defines into two files: targetconfig.h and debug.h.
For some very strange reason config.h had to be renamed targetconfig.h because no matter what when targetting the NAO it was never included properly. I suspect because Aldebaran use a file with the same name it was causing problems.

The new debug.h simply directs debug information to a file on platforms where cout is not the best place.